Enable configuration data to be read from flash (#13)
This commit adds the ability for configuration data to be read from the
pico's flash memory at start-up rather than requiring it to be compile 
time constants.

This is handled through using pointer constants based on the logic
outlined here: https://forums.raspberrypi.com//viewtopic.php?f=145&t=300146.

Additionally, this commit adds a python script to load the configuration
data into the pico. The loaded configuration data is effectively permanent,
only needing to be reloaded in the event it changes.