Improve wireless capabilities (#19)
This commit updates the implementation of Wifi to support
the ability to set a `hostname` for a device. By default, this
will be the MQTT device name for simplicity; however, the
`WifiConnection` class will accept any string.

Additionally, this commit fixes a minor bug noticed after
power outages wherein the pico and networking hardware
come up at close to the same time, resulting in the pico
needing to be rebooted. This is addressed by adding the
ability to `reset` the wireless connection.